requirements.txt
Flask==2.3.6
requests==2.32.0
beautifulsoup4==4.12.2
python-telegram-bot==20.5
